Retrieve bookdown chapters for a repository

Given an repository on GitHub, retrieve all the bookdown chapter information for the Github page if it exists. Currently only public repositories are supported.

Usage,

get_chapters(
  repo_name,
  git_pat = NULL,
  retrieve_learning_obj = FALSE,
  retrieve_keywords = TRUE,
  verbose = TRUE,
  udmodel = NULL
)

Arguments

repo_name

The full name of the repo to get bookdown chapters from. e.g. "jhudsl/OTTR_Template"

git_pat

If private repositories are to be retrieved, a github personal access token needs to be supplied. If none is supplied, then this will attempt to grab from a git pat set in the environment with usethis::create_github_token(). Authorization handled by get_git_auth

retrieve_learning_obj

TRUE/FALSE attempt to retrieve learning objectives?

retrieve_keywords

TRUE/FALSE attempt to retrieve keywords from the chapter?

verbose

TRUE/FALSE do you want more progress messages?

udmodel

A udmodel passed in for keyword determination. Will be obtained using udpipe::udpipe_download_model(language = "english") if its not given.

Value

a data frame with the repository with the following columns: data_level, data_path, chapt_name, url, repository name

Examples

if (FALSE) {

usethis::create_github_token()

get_chapters("jhudsl/Documentation_and_Usability")

}